7f5f51b6 Kovri: replace glic endian conversion with boost.endian (anonimal)
944264dc Kovri: finish the removal of i2pd's endian impl (anonimal)
7a12f4ee Kovri: implement big/native-endian bytestream write (anonimal)
dea57f7d ByteStream: implement optional endian conversion (anonimal)
b988ea7a InputByteStream: add docs to readers (anonimal)
9a301888 OutputByteStream: add write wrapper + static writer (anonimal)
ffb2fff5 Kovri: implement native-endian bytestream read (anonimal)
28099728 ByteStream: simplify boost address to vector impl (anonimal)
df947280 ByteStream: virtual dtors for I/O base cases + explicit ctors (anonimal)
889f3edb ByteStream: implement I/O base class + related I/O refactor (anonimal)
3d704a1c ClientDestination: implement bytestream in Data message handler (anonimal)
f0c8acd1 InputByteStream: add read wrapper + static reader (anonimal)
65cdc733 InputByteStream: add new getters/members (anonimal)
7900b7e3 ByteStream: cpplint/clang-format/style check-in (anonimal)
5d421dd2 ByteStream: fix copyright dates (anonimal)

+++ src/core/router/identity.h
@@ -1,5 +1,5 @@
 /**                                                                                           //
- * Copyright (c) 2013-2017, The Kovri I2P Router Project                                      //
+ * Copyright (c) 2013-2018, The Kovri I2P Router Project                                      //
  *                                                                                            //
  * All rights reserved.                                                                       //
  *                                                                                            //
@@ -296,7 +296,7 @@ class IdentityEx {
   Identity m_StandardIdentity;
   IdentHash m_IdentHash;
   mutable std::unique_ptr<kovri::core::Verifier> m_Verifier;
-  std::size_t m_ExtendedLen;
+  std::uint16_t m_ExtendedLen;
   std::unique_ptr<std::uint8_t[]> m_ExtendedBuffer;
   core::Exception m_Exception;
 };
